+++ src/main/kotlin/nl/deschepers/laraveltinker/toolwindow/TinkerOutputToolWindowFactory.kt
@@ -1,5 +1,6 @@
 package nl.deschepers.laraveltinker.toolwindow
 
+import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
@@ -12,6 +13,14 @@ class TinkerOutputToolWindowFactory : ToolWindowFactory {
     }
 
     override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
+        val actionManager = ActionManager.getInstance()
+        toolWindow.setTitleActions(
+            listOf(
+                actionManager.getAction("nl.deschepers.laraveltinker.action.OpenNewConsoleAction"),
+                actionManager.getAction("nl.deschepers.laraveltinker.action.OpenLastConsoleAction")
+            )
+        )
+
         val instance = TinkerOutputToolwindow(toolWindow)
         tinkerOutputToolWindow[project] = instance
         val contentFactory = ContentFactory.SERVICE.getInstance()