+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- [[#112]](https://github.com/Roboroads/laravel-tinker/pull/112) Adds "Open new console" and "Reopen last console" actions to the tinker output toolwindow.
+
+### Changed
+
+- Cleaned up the output toolwindow a little bit.
+
 ## [2.2.2]
 
 ### Added